refactor(agents-desktop): extract cloud networking helpers
Move cloud auth header injection and desktop server-fetch validation behind focused modules while keeping main.ts in control of current app state.
